Updated recipe for building development environment
The old recipe did not separate development packages from application
requirements. As such, the `requirements-dev.txt` file contained the
list of all dependencies.

The current implementation pipes the output of `pip freeze` command to
the `sed` command which deletes all the lines from the beginning of
the stream up to the line that matches `## The following`, and prints
the rest of the lines to the output which is then redirected to the
`requirements-dev.txt` file.